refine: move ngm client to util
